Removed unsave unwrap() methodes (and formated the code)
The code had one `expect` and one `unwrap` methode, these cause the
program to panic. In most application you do not want the application to
panic, but just handle the errors. This is why I removed those code
parts and replaced it with a `match`. Since the function already
returned an `Option` Type I made the `match` condition to exit the
function with `None`.

I also formatted the code with `rustfmt` to match the official rust
style guidelines.